% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catexplore1.R
\name{cat_explore1}
\alias{cat_explore1}
\title{cat_explore1: exploratory analyses with one categorical variable}
\usage{
cat_explore1(data, cont, cat1, alpha = 0.05, remove.na = TRUE)
}
\arguments{
\item{data}{The data frame containing the continuous response variable and categorical explanatory variable of interest -> data is an intuitive name for the data set being analyzed by thus function}

\item{cont}{The data frame column containing the continuous response variable -> cont is as an abbreviation for continuous, the type of variable this argument specifies}

\item{cat1}{The data frame column containing the categorical explanatory variable -> cat1 is as an abbreviation for continuous, the type of variable this argument specifies, plus a 1 to signify it is the first (and in this case, only) categorical variable being used for grouping}

\item{alpha}{The alpha value used to calculate the confidence interval used in plotting. Default is 0.05, producing a 95\% confidence interval.}

\item{remove.na}{Logical argument specifying if summary statistics should remove na values for their calculations. Default is TRUE.}
}
\value{
A list containing 3 plots and 2 tibbles:
\itemize{
\item {viz_hist}  is a figure with the histogram of the continuous variable
\item {median_max_min} is a tibble with the median, maximum, minimum, and range of the continuous variable grouped by the categorical variable
\item {viz_box} is a figure with boxplots of the continuous variable for each level categorical variable
\item {mean_se_n} is a tibble with the mean, standard deviation, standard error of the mean, and sample size of the continuous variable grouped by the categorical variable
\item {viz_conf} is a figure with plots of the mean and approximated confidence intervals for each level of the categorical variable.
}
}
\description{
cat_explore1 is designed for quick exploratory analyses of data frames with a continuous response variable and one explanatory categorical variable of interest. It produces plots and summary tables for investigating data distribution and comparing sample size, spread, and central tendency between levels of the categorical variable.
}
\examples{
cat_explore1(gapminder::gapminder, lifeExp, continent)
cat_explore1(gapminder::gapminder, lifeExp, continent, alpha = 0.1)
}
